Move DatabasePool to own file
